Conflicts:
	ant/libmoai/jni/Android.mk
	ant/libmoai/jni/curl/Android.mk
	samples/anim/chase/main.lua
	samples/input/input-picking/main.lua
	scons/src/NaClHost.cpp
	src/aku/AKU-iphone.mm
	src/aku/AKU.h
	src/moaicore/MOAIAction.cpp
	src/moaicore/MOAIAction.h
	src/moaicore/MOAICoroutine.cpp
	src/moaicore/MOAIFont.cpp
	src/moaicore/MOAIFont.h
	src/moaicore/MOAIGfxDevice.cpp
	src/moaicore/MOAIGfxDevice.h
	src/moaicore/MOAIInputMgr.cpp
	src/moaicore/MOAILuaObject.cpp
	src/moaicore/MOAILuaState.cpp
	src/moaicore/MOAINode.cpp
	src/moaicore/MOAIParticleEmitter.cpp
	src/moaicore/MOAIParticlePlugin.cpp
	src/moaicore/MOAIParticlePlugin.h
	src/moaicore/MOAIParticleScript.cpp
	src/moaicore/MOAIParticleScript.h
	src/moaicore/MOAIParticleState.cpp
	src/moaicore/MOAIParticleSystem.cpp
	src/moaicore/MOAIParticleSystem.h
	src/moaicore/MOAIPartitionResultBuffer.cpp
	src/moaicore/MOAIProp.h
	src/moaicore/MOAIShader.cpp
	src/moaicore/MOAIShader.h
	src/moaicore/MOAISim.cpp
	src/moaicore/MOAISim.h
	src/moaicore/MOAIStretchPatch2D.h
	src/moaicore/MOAITextBox.cpp
	src/moaicore/MOAITextBundle.cpp
	src/moaicore/MOAITextBundle.h
	src/moaicore/MOAITexture.cpp
	src/moaicore/MOAITexture.h
	src/moaicore/MOAITransform.cpp
	src/moaicore/MOAITransform.h
	src/moaicore/moaicore.cpp
	src/moaicore/moaicore.h
	src/moaicore/shaders/MOAIMeshShader-vsh.h
	src/moaiext-iphone/moaiext-iphone.h
	src/uslscore/USBinarySearch.h
	src/uslscore/USFileSys.cpp
	src/uslscore/USParser.cpp
	src/zipfs/ZIPFSZipFile.c
	src/zipfs/zipfs.c
	vs2010/moaicore/moaicore.vcxproj
	vs2010/moaicore/moaicore.vcxproj.filters
	xcode/ios/MoaiSample.xcodeproj/project.pbxproj
	xcode/ios/MoaiSample.xcodeproj/xcshareddata/xcschemes/moai.xcscheme
	xcode/libmoai/libmoai.xcodeproj/project.pbxproj

+++ src/moaiext-audiosampler/MOAIAudioSampler.cpp
@@ -122,7 +122,7 @@ void MOAIAudioSampler::globalCallback( void *inUserData,
 int	MOAIAudioSampler::_prepareBuffer ( lua_State* L ) {
     MOAI_LUA_SETUP ( MOAIAudioSampler, "UNN" )
         ;
-    double sec = state.GetValue < float > (2, 1);
+    float sec = state.GetValue < float > (2, 1);
     self->mBufferAryLen = state.GetValue < u32> (3, 5 );
 
     memset( &self->recFmt, 0, sizeof(self->recFmt));
